\name{accrual.gui}
\alias{accrual.gui}
\title{
GUI Version of the Bayesian Accrual Prediciton 
}
\description{
The R GUI interface only needs the researchers to input the original design information that are required information for IRBs (total time proposed and total subjects proposed) and the updated accrual data (time since start and subjects accrual). It uses Bayesian prediction model in the background of calculation }

\usage{
accrual.gui()
}
\author{
Yu Jiang, Steve Simon, Matthew S. Mayo , Rama Raghavan, Byron J. Gajewski
}
\examples{
accrual.gui()
}

\keyword{gui }
\keyword{Bayesian}
