\name{accrual.gui}
\alias{accrual.gui}
\title{
GUI Version of Bayesian Accrual Prediction
}
\description{
The Graphical User Interface (GUI) needs the information from the original design of protocol (e.g., total enrollment time proposed and total participants proposed) and the ongoing accrual data (e.g., enrollment time since start and number of participants enrolled). The Bayesian prediction model is implemented in the background of calculation.
}
\usage{
accrual.gui()
}
\arguments{
There are no arguments for this function.
}
\value{
A list of prediction on number of participants that recruited in a fixed time frame or on time frame that needed for reaching a certain number of participants. A Single-center or Multi-center selection is available for number of participants prediction.
}
\author{
Junhao Liu, Yu Jiang, Cen Wu, Steve Simon, Matthew S. Mayo, Rama Raghavan, Byron J. Gajewski
}
\examples{
accrual.gui()
}

\keyword{gui}
\keyword{Bayesian}
